backend/requirements.txt
apscheduler==3.9.1
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.27
mysql-connector-python==8.2.0
PyMySQL==1.1.0
alembic==1.13.1
bcrypt==4.0.1
httpx==0.24.1
python-dotenv==0.21.0
python-jose==3.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
fastapi-mail==1.2.9
pandas==2.2.3

